NtupleAnalysis DataFormat: changed fNUP,fNPU,fSimulatedNPU from short to int to remove warnings on screen

+++ NtupleAnalysis/src/DataFormat/interface/EventID.h
@@ -32,7 +32,7 @@ private:
   const Branch<unsigned long long> *fEvent;
   const Branch<unsigned int> *fLumi;
   const Branch<unsigned int> *fRun;
-  const Branch<short> *fNUP;
+  const Branch<int> *fNUP;
   const Branch<float> *fPrescale;
 };
